S2 Plume finder cross-sectional flux progress
Added a function that takes a centroid to the plume polygons and draws a line perpendicular to the plume direction, the width of the plume. Then a bresenham's line algorithm is used to find pixels that intersect with that line. These pixels are then summed. Finally the median dataset value is multiplied by the number of cross sectional pixels and this figure is subtracted from the summed cross sectional value to leave the digital number above background value.