Regression analysis added
In an attempt to draw emission rate from cross sectional flux and wind speed a regression analysis was performed and a model created to estimate detected plume emission rates. Additionally the Hassi_Messaoud_urban.geojson is for masking the urban areas so that they don't interfere in the analysis but this hasn't been completed yet.