Added urban masking feature
The urban areas of Hassi Messaoud were causing problems with the plume cross section summation when the plumes went over those areas. The urban area .geojson was used as the basis to render the urban pixels with a value of zero. The other fields need to be surveyed for towns (there aren't many) too.