Reworked code for plume finder and added IME measurement tool (work in progress)
Discovered an error in how Varon et al's (2021) MBMP method was being implemented. This has almost been fixed. One last thing that needs to be changed is the SWIR MBMP dataset needs to be forced to only have positive values for the cross sectional sum maths to work. I'll sort that out tomorrow.

The IME measurement tool produces data but it doesn't match examples in peer reviewed sources. Hopefully I will be able to fix it.