Updates to plume tagging
Finally managed to get the plume tagging system to work on an area rather than a point. This means the user doesn't need to worry about precision when tagging as they can be off the mark by 20 pixels and it will still tag the plume.

This now means that monitoring of known emitters is now possible by default. 11 known emitters have been added to the initial_data code. These should not be touched by the end user so that those locations can always be checked for plumes.

Finally the How to Guide has been upfated to reflect the new code.