Replaced Random forest model for XGBoost
The Random forest was producing a couple of bad errors when low wind was combined with high plume length and vice versa. After a lot of trying to fix this by trying to factor in wind speed to length, it was decided to change the algorithm for prediction. Extreme Gradient Boosting was selected after testing multiple models for their predictability.