Updated CH4 Sentinel for public use.
The Specific Algerian oil fields have been replaced with a system that allows the end user to select a lat long in decimal degrees and set a specific bounding box in decimal degrees. Once entered the code returns the name of the place to avoid confusion.

The end data frame now gives results in tonnes per hour of methane.

The known point sources for Hassi Messaoud have been removed.

Updated the environment file to include Geopy